move word to query
